+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 
 ## [Unreleased]
 ### Added
+- Check that Microsoft.VisualStudio.Threading is not used.
 ### Fixed
 ### Changed
 - FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.11.0